refactor(ccs): Separate lib.rs into modules
